[refactor] 멤버 Entity 필드 수정 #77
