--- src/main/java/classfit/example/classfit/member/domain/Gender.java
@@ -1,5 +1,0 @@
-package classfit.example.classfit.member.domain;
-
-public enum Gender {
-    MALE, FEMALE, NONE
-}
