--- src/main/java/classfit/example/classfit/member/domain/LoginType.java
@@ -1,5 +1,0 @@
-package classfit.example.classfit.member.domain;
-
-public enum LoginType {
-    KAKAO, NAVER, GOOGLE
-}
