+++ src/main/java/classfit/example/classfit/member/domain/Member.java
@@ -1,5 +1,6 @@
 package classfit.example.classfit.member.domain;
 
+import classfit.example.classfit.academy.domain.Academy;
 import classfit.example.classfit.category.domain.MainClass;
 import classfit.example.classfit.common.domain.BaseEntity;
 import jakarta.persistence.*;
@@ -7,7 +8,6 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @Entity
@@ -20,28 +20,11 @@ public class Member extends BaseEntity {
     @Column(name = "member_id")
     private Long id;
 
-    @Column(nullable = false)
-    private Long socialId;
-
     @Column(nullable = false, length = 100)
     private String email;
 
     @Column(nullable = false, length = 30)
     private String name;
-
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "VARCHAR(10)", nullable = false)
-    private Gender gender;
-
-    @Column(nullable = false, length = 10)
-    private LocalDate birth;
-
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "VARCHAR(20)", nullable = false)
-    private LoginType loginType;
-
-    @Column(nullable = false, length = 30)
-    private String department;
 
     @Column(nullable = false, length = 20)
     private String phoneNumber;
@@ -52,5 +35,9 @@ public class Member extends BaseEntity {
 
     @OneToMany(mappedBy = "member", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<MainClass> mainClasses;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "academy_id", nullable = false)
+    private Academy academy;
 
 }
