+++ src/main/java/classfit/example/classfit/student/domain/Gender.java
@@ -1,0 +1,5 @@
+package classfit.example.classfit.student.domain;
+
+public enum Gender {
+    MALE, FEMALE, NONE
+}
