+++ src/main/java/classfit/example/classfit/student/dto/request/StudentRequest.java
@@ -1,7 +1,7 @@
 package classfit.example.classfit.student.dto.request;
 
 import classfit.example.classfit.common.validation.EnumValue;
-import classfit.example.classfit.member.domain.Gender;
+import classfit.example.classfit.student.domain.Gender;
 import classfit.example.classfit.student.domain.Student;
 import jakarta.validation.constraints.*;
 import org.springframework.format.annotation.DateTimeFormat;
