+++ src/main/java/classfit/example/classfit/student/service/StudentService.java
@@ -8,7 +8,7 @@ import classfit.example.classfit.category.repository.SubClassRepository;
 import classfit.example.classfit.classStudent.domain.ClassStudent;
 import classfit.example.classfit.classStudent.repository.ClassStudentRepository;
 import classfit.example.classfit.common.exception.ClassfitException;
-import classfit.example.classfit.member.domain.Gender;
+import classfit.example.classfit.student.domain.Gender;
 import classfit.example.classfit.student.domain.Student;
 import classfit.example.classfit.student.dto.request.StudentRequest;
 import classfit.example.classfit.student.dto.request.StudentUpdateRequest;
